+++ packages/base/sync/__tests__/SyncManager.spec.ts
@@ -1013,3 +1013,321 @@ it('should reset if syncmanager snapshot and collection are not in sync', async 
   // @ts-expect-error - private property
   expect(syncManager.remoteChanges.length).toBe(0)
 })
+
+it('should start sync after internal collections are ready', async () => {
+  const persistenceAdapter = memoryPersistenceAdapter([], undefined, 100)
+  const mockPersistenceAdapter = createPersistenceAdapter({
+    register: vi.fn(persistenceAdapter.register),
+    load: vi.fn(persistenceAdapter.load),
+    save: vi.fn(persistenceAdapter.save),
+  })
+  const mockPull = vi.fn<() => Promise<LoadResponse<TestItem>>>().mockResolvedValue({
+    items: [{ id: '1', name: 'Test Item' }],
+  })
+
+  const mockPush = vi.fn<(options: any, pushParameters: any) => Promise<void>>()
+    .mockResolvedValue()
+
+  const syncManager = new SyncManager({
+    persistenceAdapter: () => mockPersistenceAdapter,
+    pull: mockPull,
+    push: mockPush,
+  })
+
+  let persistenceInitialized = false
+  void Promise.all([
+    new Promise((resolve) => {
+      // @ts-expect-error - private property
+      syncManager.syncOperations.once('persistence.init', resolve)
+    }),
+    new Promise((resolve) => {
+      // @ts-expect-error - private property
+      syncManager.changes.once('persistence.init', resolve)
+    }),
+    new Promise((resolve) => {
+      // @ts-expect-error - private property
+      syncManager.snapshots.once('persistence.init', resolve)
+    }),
+  ]).then(() => {
+    persistenceInitialized = true
+  })
+
+  const collection = new Collection<TestItem, string, any>()
+  syncManager.addCollection(collection, { name: 'test' })
+
+  expect(mockPersistenceAdapter.load).not.toBeCalled()
+  expect(mockPull).not.toBeCalled()
+  expect(persistenceInitialized).toBeFalsy()
+  await syncManager.sync('test')
+
+  expect(mockPull).toBeCalled()
+  expect(mockPersistenceAdapter.load).toHaveBeenCalledBefore(mockPull)
+  expect(persistenceInitialized).toBeTruthy()
+})
+
+it('should start sync after collection is ready', async () => {
+  const persistenceAdapter = memoryPersistenceAdapter([], undefined, 100)
+  const mockPersistenceAdapter = createPersistenceAdapter({
+    register: vi.fn(persistenceAdapter.register),
+    load: vi.fn(persistenceAdapter.load),
+    save: vi.fn(persistenceAdapter.save),
+  })
+  const mockPull = vi.fn<() => Promise<LoadResponse<TestItem>>>().mockResolvedValue({
+    items: [{ id: '1', name: 'Test Item' }],
+  })
+
+  const mockPush = vi.fn<(options: any, pushParameters: any) => Promise<void>>()
+    .mockResolvedValue()
+
+  const syncManager = new SyncManager({
+    pull: mockPull,
+    push: mockPush,
+  })
+
+  const collection = new Collection<TestItem, string, any>({
+    persistence: mockPersistenceAdapter,
+  })
+  let persistenceInitialized = false
+  void new Promise<void>((resolve) => {
+    collection.once('persistence.init', resolve)
+  }).then(() => {
+    persistenceInitialized = true
+  })
+
+  syncManager.addCollection(collection, { name: 'test' })
+
+  expect(mockPull).not.toBeCalled()
+  expect(mockPersistenceAdapter.load).not.toBeCalled()
+  expect(persistenceInitialized).toBeFalsy()
+  await syncManager.sync('test')
+
+  expect(mockPull).toBeCalled()
+  expect(mockPersistenceAdapter.load).toHaveBeenCalledBefore(mockPull)
+  expect(persistenceInitialized).toBeTruthy()
+})
+
+it('should fail if there was a persistence error during initialization', async () => {
+  const persistenceAdapter = memoryPersistenceAdapter([], undefined, 100)
+  const mockPersistenceAdapter = createPersistenceAdapter({
+    register: vi.fn(persistenceAdapter.register),
+    load: vi.fn(() => Promise.reject(new Error('Persistence error'))),
+    save: vi.fn(persistenceAdapter.save),
+  })
+  const mockPull = vi.fn<() => Promise<LoadResponse<TestItem>>>().mockResolvedValue({
+    items: [{ id: '1', name: 'Test Item' }],
+  })
+
+  const mockPush = vi.fn<(options: any, pushParameters: any) => Promise<void>>()
+    .mockResolvedValue()
+
+  const syncManager = new SyncManager({
+    pull: mockPull,
+    push: mockPush,
+  })
+
+  const collection = new Collection<TestItem, string, any>({
+    persistence: mockPersistenceAdapter,
+  })
+  let persistenceInitialized = false
+  void new Promise<void>((resolve) => {
+    collection.once('persistence.init', resolve)
+  }).then(() => {
+    persistenceInitialized = true
+  })
+  let persistenceError = false
+  void new Promise<void>((resolve) => {
+    collection.once('persistence.error', () => resolve())
+  }).then(() => {
+    persistenceError = true
+  })
+
+  syncManager.addCollection(collection, { name: 'test' })
+
+  expect(mockPull).not.toBeCalled()
+  expect(persistenceInitialized).toBeFalsy()
+  expect(persistenceError).toBeFalsy()
+
+  await expect(syncManager.sync('test')).rejects.toThrowError('Persistence error')
+
+  expect(mockPull).not.toBeCalled()
+  expect(persistenceInitialized).toBeFalsy()
+  expect(persistenceError).toBeTruthy()
+})
+
+it('should start sync if autostart is enabled', async () => {
+  const mockPull = vi.fn<() => Promise<LoadResponse<TestItem>>>().mockResolvedValue({
+    items: [{ id: '1', name: 'Test Item' }],
+  })
+  const mockPush = vi.fn<(options: any, pushParameters: any) => Promise<void>>()
+    .mockResolvedValue()
+
+  const registerRemoteChange = vi.fn()
+  const syncManager = new SyncManager({
+    pull: mockPull,
+    push: mockPush,
+    registerRemoteChange,
+    autostart: true,
+  })
+
+  const collection = new Collection<TestItem, string, any>()
+  syncManager.addCollection(collection, { name: 'test' })
+
+  expect(registerRemoteChange).toBeCalledTimes(1)
+  await syncManager.sync('test')
+  expect(mockPull).toBeCalled()
+
+  await syncManager.startSync('test')
+  expect(registerRemoteChange).toBeCalledTimes(1) // already started, should not be called again
+})
+
+it('should not start sync if autostart is disabled', async () => {
+  const mockPull = vi.fn<() => Promise<LoadResponse<TestItem>>>().mockResolvedValue({
+    items: [{ id: '1', name: 'Test Item' }],
+  })
+  const mockPush = vi.fn<(options: any, pushParameters: any) => Promise<void>>()
+    .mockResolvedValue()
+
+  const registerRemoteChange = vi.fn()
+  const syncManager = new SyncManager({
+    pull: mockPull,
+    push: mockPush,
+    registerRemoteChange,
+    autostart: false,
+  })
+
+  const collection = new Collection<TestItem, string, any>()
+  syncManager.addCollection(collection, { name: 'test' })
+
+  expect(registerRemoteChange).toBeCalledTimes(0)
+  await syncManager.sync('test')
+  expect(mockPull).toBeCalled()
+  expect(registerRemoteChange).toBeCalledTimes(0)
+
+  await syncManager.startSync('test')
+  expect(registerRemoteChange).toBeCalledTimes(1)
+})
+
+it('should not trigger sync if collection is paused', async () => {
+  const mockPull = vi.fn<() => Promise<LoadResponse<TestItem>>>().mockResolvedValue({
+    items: [{ id: '1', name: 'Test Item' }],
+  })
+  const mockPush = vi.fn<(options: any, pushParameters: any) => Promise<void>>()
+    .mockResolvedValue()
+
+  type SyncManagerOptions = ConstructorParameters<typeof SyncManager>[0]
+  type RegisterRemoteChangeParameters = Parameters<NonNullable<SyncManagerOptions['registerRemoteChange']>>
+  type CollectionOptions = RegisterRemoteChangeParameters[0]
+  type RemoteChangeHandler = RegisterRemoteChangeParameters[1]
+
+  let onRemoteChangeHandler: RemoteChangeHandler | undefined
+  const cleanupFunction = vi.fn()
+  const registerRemoteChange = vi.fn((
+    options: CollectionOptions,
+    onRemoteChange: RemoteChangeHandler,
+  ) => {
+    onRemoteChangeHandler = onRemoteChange
+    return cleanupFunction
+  })
+  const syncManager = new SyncManager({
+    pull: mockPull,
+    push: mockPush,
+    registerRemoteChange,
+    autostart: false,
+  })
+
+  const collection = new Collection<TestItem, string, any>()
+  syncManager.addCollection(collection, { name: 'test' })
+
+  expect(registerRemoteChange).toBeCalledTimes(0)
+  expect(mockPull).toBeCalledTimes(0)
+  expect(onRemoteChangeHandler).toBeUndefined()
+
+  await syncManager.startSync('test')
+  if (!onRemoteChangeHandler) return
+  expect(cleanupFunction).toBeCalledTimes(0)
+  expect(mockPull).toBeCalledTimes(0)
+  expect(registerRemoteChange).toBeCalledTimes(1)
+  expect(onRemoteChangeHandler).toBeDefined()
+
+  await onRemoteChangeHandler()
+  expect(mockPull).toBeCalledTimes(1)
+
+  await syncManager.pauseSync('test')
+  await onRemoteChangeHandler()
+  expect(cleanupFunction).toBeCalledTimes(1)
+  expect(mockPull).toBeCalledTimes(2)
+})
+
+it('should only automatically push if started', async () => {
+  const mockPull = vi.fn<() => Promise<LoadResponse<TestItem>>>().mockResolvedValue({
+    items: [{ id: '1', name: 'Test Item' }],
+  })
+
+  const mockPush = vi.fn<(options: any, pushParameters: any) => Promise<void>>()
+    .mockResolvedValue()
+  const onError = vi.fn()
+
+  const syncManager = new SyncManager({
+    onError,
+    pull: mockPull,
+    push: mockPush,
+    autostart: false,
+  })
+
+  const mockCollection = new Collection<TestItem, string, any>()
+
+  syncManager.addCollection(mockCollection, { name: 'test' })
+  mockCollection.insert({ id: '2', name: 'New Item' })
+
+  await new Promise((resolve) => {
+    setTimeout(resolve, 110)
+  })
+
+  expect(onError).not.toHaveBeenCalled()
+  expect(mockPush).toHaveBeenCalledTimes(0)
+
+  await syncManager.startSync('test')
+  await new Promise((resolve) => {
+    setTimeout(resolve, 110)
+  })
+  expect(onError).not.toHaveBeenCalled()
+  expect(mockPush).toHaveBeenCalledTimes(1)
+})
+
+it('should handle an error during registerRemoteChange', async () => {
+  const mockPull = vi.fn<() => Promise<LoadResponse<TestItem>>>().mockResolvedValue({
+    items: [{ id: '1', name: 'Test Item' }],
+  })
+  const mockPush = vi.fn<(options: any, pushParameters: any) => Promise<void>>()
+    .mockResolvedValue()
+  const onError = vi.fn()
+  const syncManager = new SyncManager({
+    pull: mockPull,
+    push: mockPush,
+    autostart: true,
+    registerRemoteChange: () => new Promise((resolve, reject) => {
+      reject(new Error('Failed to register remote change'))
+    }),
+    onError,
+  })
+
+  const collection = new Collection<TestItem, string, any>()
+  syncManager.addCollection(collection, { name: 'test' })
+  await new Promise((resolve) => {
+    setTimeout(resolve, 0)
+  })
+  expect(onError).toHaveBeenCalledTimes(1)
+})
+
+it('should return a tuple from getCollection function', () => {
+  const syncManager = new SyncManager({
+    pull: vi.fn(),
+    push: vi.fn(),
+  })
+  const collection = new Collection<TestItem, string, any>()
+  syncManager.addCollection(collection, { name: 'test' })
+  // eslint-disable-next-line @typescript-eslint/no-deprecated
+  const [collectionInstance, collectionOptions] = syncManager.getCollection('test')
+  expect(collectionInstance).toBeInstanceOf(Collection)
+  expect(collectionOptions).toEqual({ name: 'test' })
+})